Require authentication for static catalog files
