+++ api/routes/challenge.routes.js
@@ -1,10 +1,11 @@
 import {
-  verifyHeaders,
   requestChallenge,
   verifyChallenge,
 } from '../controllers/challenge.controllers.js';
 
+import { verifyNetID } from '../utils.js';
+
 export default (app) => {
-  app.get('/api/challenge/request', verifyHeaders, requestChallenge);
+  app.get('/api/challenge/request', verifyNetID, requestChallenge);
   app.post('/api/challenge/verify', verifyChallenge);
 };
