+++ api/utils.js
@@ -54,6 +54,28 @@ export function getRandomInt(max) {
 }
 
 /**
+ * Middleware to verify request headers
+ *
+ * @prop req - express request object
+ * @prop res - express response object
+ * @prop next - express next object
+ */
+export const verifyNetID = (req, res, next) => {
+  // get authentication headers
+  const netid = req.header('x-coursetable-netid'); // user's NetID
+  const authd = req.header('x-coursetable-authd'); // if user is logged in
+
+  // require NetID authentication
+  if (authd !== 'true' || !netid) {
+    return res.status(401).json({
+      error: 'NOT_AUTHENTICATED',
+    });
+  }
+
+  next();
+};
+
+/**
  * Get static catalogs for each season from Hasura,
  * @prop overwrite - whether or not to skip existing catalogs.
  */
